Rename clickquiry to cliq and remove util package
