+++ maubot/cli/cliq/__init__.py
@@ -1,0 +1,2 @@
+from .cliq import command, option
+from .validators import SPDXValidator, VersionValidator, PathValidator
